+++ src/ents/shop_tile_triggers.hpp
@@ -9,7 +9,7 @@ struct State;
 
 namespace ents::shop {
 
-StageTileTrigger MakeShopVandalismTileTrigger(const IVec2& tile_pos, std::size_t target_spawn_index);
+StageTileTrigger MakeShopVandalismTileTrigger(const IVec2& tile_pos, std::uint32_t target_spawn_index);
 StageTileTrigger MakeShopVandalismTileTrigger(const IVec2& tile_pos, VID target_vid);
 
 void OnShopVandalismTileDestroyed(
